+++ TestAutomationTool/src/main/java/ej/fp/widget/command/CommandBuilder.java
@@ -1,9 +1,8 @@
 /*
  * Java
  *
- * Copyright 2021 MicroEJ Corp. All rights reserved.
- * This library is provided in source code for use, modification and test, subject to license terms.
- * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
+ * Copyright 2021-2022 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package ej.fp.widget.command;
 
@@ -26,6 +25,8 @@ public class CommandBuilder {
 		register("compareScreenshot", CompareScreenshotCommand.class); //$NON-NLS-1$
 		register("buttonPress", ButtonCommand.Press.class); //$NON-NLS-1$
 		register("buttonRelease", ButtonCommand.Release.class); //$NON-NLS-1$
+		register("buttonLongPress", LongButtonCommand.Press.class); //$NON-NLS-1$
+		register("buttonLongRelease", LongButtonCommand.Release.class); //$NON-NLS-1$
 	}
 
 	/**