+++ TestAutomationJUnitLibrary/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - Unreleased
 
+### Changed
+
+- Change license terms to MicroEJ Corp. BSD-style license.
+
+## [1.0.0] - 2021-10-27
+
+### Changed
+
+  - First public release and packaging.
 
 ## [0.1.0] - 2021-09-30
 
@@ -15,6 +25,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---  
 _Markdown_   
-_Copyright 2021 MicroEJ Corp. All rights reserved._
-_This library is provided in source code for use, modification and test, subject to license terms._
-_Any modification of the source code will break MicroEJ Corp. warranties on the whole library._
+_Copyright 2021-2022 MicroEJ Corp. All rights reserved._
+_Use of this source code is governed by a BSD-style license that can be found with this software._